Setup HTML and basic styles
Setup the basic HTML.
Wrap the nav a href and its
dropdown (ul and li)in a div
with a
class. Also give the nav
a href and its corresponding
ul and li relevant classes.
This is so that they can be
targeted by the CSS and JS.

Add basic styling. Style the
relevant ul and its li
(.dropdown-menu) so that
their display is set to
none. Later, the JS will
chanadd a class that will
reveal their display.