Create dropdown module
Create a new dropdown
directory and a file.

Initialize and export a
Dropdown class. The class
constructor selects a
nodelist of all elements
with class of dropdown-menu
that are descendants of
elements with class of
dropdown. If it finds
the nodelist, it call
the intialize method.

For now, the initialize
method
just places a click event
listener on the document.

Import the Dropdown class
into the index.js and
create a new instance of
it.