Handle dropdown element clicks
Initialize a handleClicks()
method that takes in a
dropdownActionElement
as an argument. The
method loops through the
dropdowns (menus) defined in the
constructor and checks each, if
its parent element contains
dropdownActionElement. If it does,
its adds the active class to its
class list.

In initialize(), check whether the
element that was clicked on
has the class - dropdown-action
and if it does, call handleClicks()
and pass in the element as an
argument.

This effectively checks whether
dropdownActionElement is contained
in the parent element of dropdowns.
Clicking on any of these links with
a dropdown menu reveals the dropdown
menu.