Conditionally hide dropdown menu
Add an else block that removes
the active class from any
dropdown which its parent element
does not contain the the
href element (.dropdown-action )
that is clicked on.

This effectively makes it that
when one dropdown-action is
clicked, its dropdown-menu is
visible and any other
dropdown-action's dropdown-menu
is hidden.

Update handleClicks parameter
name to camelCase.