Hide dropdown on click outside
Define a method
hideDropdownOnClickOutside that
takes in a an clickedElement as an
argument. It loops through the
the dropdowns defined in the
constructor and for each dropdown,
it removes the
active class from the dropdown
if the parent element of the
dropdown does not contain
the clickedElement.

This will hide a dropdown
that is active when any part of the
document (that is not another
dropdown action href) is clicked.

Call this method with the
event.target argument inside a
new else block in the
initialize() method.

Thus, if dropdown action is
clicked show dropdown, if
another dropdown action is
clicked next, show dropdown.
If another part of the
document is clicked when
dropdown is active, hide
the active dropdown.