Handle clicks inside dropdown
Clicks outside .dropdown
was effectively calling
hideDropdownOnClickOutside()
and hiding any active dropdown
menus. hideDropdownOnClickOutside()
checks if the click target is
outside the .dropdown container.

Clicking on the
dropdown menu li a items
is still clicking inside
the .dropdown container,
so the check in
hideDropdownOnClickOutside()
fails, and the dropdown
doesn’t close.

Update hideDropdownOnClickOutside()
to take in a default parameter
of null. Call it without a parameter
in the else block of initialize.
contains(null) always evaluates to
false and ! makes it true in
hideDropdownOnClickOutside().
Calling it in initialize()
this way ensures that whenever
a click happens on any element
that doesn’t match .dropdown-action,
hideDropdownOnClickOutside()
is triggered with no arguments
effectively hiding all dropdowns.