Update package.json and README
Update the name, version,
and author.

Update the README.md

Add a .npmignore file as
the package was pushed on
npmjs.com

Prettier changed the
alignment of code in
index.html and index.css

In the index.js, create a
new instance of Dropdown
after the whole page
has been loaded.